[Refactor] LottoGame 클래스 생성 및 로직 이동
